Fixed Portfolio
Fixed issues with backgrounds, Made the dots not run into the slider on
short screens; added a max-height media query to change the font size of
the header. added text-shadow, which looks better than expected. Added
some shadow to the blurb text on the index, not sure how I feel about it